+++ src/components/content/zus-aggregated-profile/zus-aggregated-profile-tabs.tsx
@@ -39,6 +39,15 @@ import { ZAPResourceName } from "@/components/content/zus-aggregated-profile/zus
 import { TabGroupItem } from "@/components/core/tab-group/tab-group";
 import i18next from "@/i18n";
 import { PatientMedicationsOutsideBadge } from "../medications/patient-medications-outside-badge";
+import {
+  PatientObservationsProps,
+  PatientObservations,
+} from "../observations/patient-observations";
+import {
+  PatientObservationsOutsideProps,
+  PatientObservationsOutsideBadge,
+  PatientObservationsOutside,
+} from "../observations/patient-observations-outside";
 
 export type ZusAggregatedProfileTabs = Record<
   ZAPResourceName,
@@ -47,7 +56,7 @@ export type ZusAggregatedProfileTabs = Record<
 
 export const zusAggregatedProfileTabs: ZusAggregatedProfileTabs = {
   allergies: (props: PatientAllergiesProps = {}) => ({
-    key: "allergies-builder-records",
+    key: "allergies",
     getPanelClassName: () => "ctw-pt-5",
     display: () => "allergy list",
     render: () => <PatientAllergies {...props} />,
@@ -60,13 +69,13 @@ export const zusAggregatedProfileTabs: ZusAggregatedProfileTabs = {
   }),
 
   conditions: (props: PatientConditionsProps = {}) => ({
-    key: "condition-provider-records",
+    key: "conditions",
     display: () => i18next.t("zap.tabs.conditions"),
     render: () => <PatientConditions {...props} />,
   }),
 
   "conditions-outside": (props: PatientConditionsOutsideProps = {}) => ({
-    key: "condition-outside-records",
+    key: "conditions-outside",
     display: () => (
       <div className="ctw-space-x-2">
         <span className="ctw-capitalize">
@@ -86,20 +95,20 @@ export const zusAggregatedProfileTabs: ZusAggregatedProfileTabs = {
   }),
 
   immunizations: (props: PatientImmunizationsProps = {}) => ({
-    key: "immunization-outside-records",
+    key: "immunizations",
     getPanelClassName: () => "ctw-pt-5",
     display: () => "immunizations",
     render: () => <PatientImmunizations {...props} />,
   }),
 
   medications: (props: PatientMedicationsProps = {}) => ({
-    key: "medication-builder-records",
+    key: "medications",
     display: () => "medication list",
     render: () => <PatientMedications {...props} />,
   }),
 
   "medications-outside": (props: PatientMedicationsOutsideProps = {}) => ({
-    key: "other-provider-records",
+    key: "medications-outside",
     display: () => (
       <div className="ctw-space-x-2">
         <span className="ctw-capitalize">outside medications</span>
@@ -107,6 +116,23 @@ export const zusAggregatedProfileTabs: ZusAggregatedProfileTabs = {
       </div>
     ),
     render: () => <PatientMedicationsOutside {...props} />,
+  }),
+
+  observations: (props: PatientObservationsProps = {}) => ({
+    key: "observations",
+    display: () => "observations",
+    render: () => <PatientObservations {...props} />,
+  }),
+
+  "observations-outside": (props: PatientObservationsOutsideProps = {}) => ({
+    key: "observations-outside",
+    display: () => (
+      <>
+        <span className="ctw-pr-2 ctw-capitalize">outside observations</span>
+        <PatientObservationsOutsideBadge />
+      </>
+    ),
+    render: () => <PatientObservationsOutside {...props} />,
   }),
 
   timelines: (props: PatientTimelineProps = {}) => ({