+++ src/index.ts
@@ -7,6 +7,7 @@ export * from "@/components/content/immunizations/patient-immunizations";
 export * from "@/components/content/care-team/patient-careteam";
 export * from "@/components/content/document/patient-documents";
 export * from "@/components/content/medications/patient-medications";
+export * from "@/components/content/observations/patient-observations";
 export * from "@/components/content/patient-history/use-patient-history";
 export * from "@/components/content/patients/patients-table";
 export * from "@/components/content/patients/patients-search";